+++ data-raw/creating_unga_tsv.R
@@ -5,4 +5,4 @@ library(data.table)
 
 CD <- CorpusData$new()
 CD$import_xml(filenames = unga_xml_files)
-data.table::fwrite(x = CD$chunktable, file = "~/Lab/github/bignlp/inst/extdata/tsv/unga.tsv")
+data.table::fwrite(x = CD$chunktable, file = "~/Lab/github/bignlp/inst/extdata/tsv/unga.tsv", sep = "\t")